CHORE Partially reworked project structure
