+++ ptero_common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ptero_common"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bitvec = "0.22"
+rand = "0.8"
+derive_builder = "0.10"
+snafu = "0.6"
