+++ ptero_derive/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ptero_derive"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+name = "ptero_derive"
+proc-macro = true
+
+[dependencies]
+syn = "1.0"
+quote = "1.0"
+ptero_common = { version = "0.1", path = "../ptero_common" }
