+++ ptero_text/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ptero_text"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4"
+rand = "0.8"
+bitvec = "0.22"
+derive_builder = "0.10"
+unicode-segmentation = "1.8"
+snafu = "0.6"
+ptero_common = { version = "0.1", path = "../ptero_common" }
