+++ src/method.rs
@@ -27,10 +27,6 @@ pub mod trailing_unicode;
 pub mod complex;
 
 pub mod variant;
-
-pub mod config;
-pub mod extended_line_method;
-
 /// Combination of [Encoder](crate::encoder::Encoder) and [Decoder](crate::decoder::Decoder) traits - each method should be able to encode and decode.
 pub trait Method<E, D>: Encoder<E> + Decoder<D>
 where
@@ -38,22 +34,4 @@ where
     D: Context,
 {
     fn method_name(&self) -> String;
-}
-
-pub trait SteganographyMethod<Cover, Err> {
-    type ConcealedOutput;
-
-    fn try_conceal<Order, Type>(
-        &mut self,
-        cover: Cover,
-        data: &mut Iter<Order, Type>,
-    ) -> Result<Self::ConcealedOutput, Err>
-    where
-        Order: BitOrder,
-        Type: BitStore;
-
-    fn try_reveal<Order, Type>(&mut self, cover: Cover) -> Result<BitVec<Order, Type>, Err>
-    where
-        Order: BitOrder,
-        Type: BitStore;
 }
